+++ Platform/OpenCore/OpenCoreMisc.c
@@ -20,9 +20,11 @@ WITHOUT WARRANTIES OR REPRESENTATIONS OF ANY KIND, EITHER EXPRESS OR IMPLIED.
 #include <Library/DebugLib.h>
 #include <Library/DevicePathLib.h>
 #include <Library/MemoryAllocationLib.h>
+#include <Library/OcAcpiLib.h>
 #include <Library/OcAppleBootPolicyLib.h>
 #include <Library/OcConsoleLib.h>
 #include <Library/OcDebugLogLib.h>
+#include <Library/OcSmbiosLib.h>
 #include <Library/OcStringLib.h>
 #include <Library/PrintLib.h>
 #include <Library/UefiBootServicesTableLib.h>
@@ -70,6 +72,92 @@ OcStoreLoadPath (
     OutPath,
     Status
     ));
+}
+
+STATIC
+EFI_STATUS
+ProduceDebugReport (
+  IN EFI_HANDLE  VolumeHandle
+  )
+{
+  EFI_STATUS         Status;
+  EFI_FILE_PROTOCOL  *Fs;
+  EFI_FILE_PROTOCOL  *SysReport;
+  EFI_FILE_PROTOCOL  *SubReport;
+
+  if (VolumeHandle != NULL) {
+    Fs = LocateRootVolume (VolumeHandle, NULL);
+  } else {
+    Fs = NULL;
+  }
+
+  if (Fs == NULL) {
+    Status = FindWritableFileSystem (&Fs);
+    if (EFI_ERROR (Status)) {
+      DEBUG ((DEBUG_INFO, "OC: No usable filesystem for report - %r\n", Status));
+      return EFI_NOT_FOUND;
+    }
+  }
+
+  Status = SafeFileOpen (
+    Fs,
+    &SysReport,
+    L"SysReport",
+    EFI_FILE_MODE_READ | EFI_FILE_MODE_WRITE,
+    EFI_FILE_DIRECTORY
+    );
+  if (!EFI_ERROR (Status)) {
+    DEBUG ((DEBUG_INFO, "OC: Report is already created, skipping\n"));
+    SysReport->Close (SysReport);
+    Fs->Close (Fs);
+    return EFI_ALREADY_STARTED;
+  }
+
+  Status = SafeFileOpen (
+    Fs,
+    &SysReport,
+    L"SysReport",
+    EFI_FILE_MODE_READ | EFI_FILE_MODE_WRITE | EFI_FILE_MODE_CREATE,
+    EFI_FILE_DIRECTORY
+    );
+  if (EFI_ERROR (Status)) {
+    DEBUG ((DEBUG_INFO, "OC: Cannot create SysReport - %r\n", Status));
+    Fs->Close (Fs);
+    return Status;
+  }
+
+  Status = SafeFileOpen (
+    SysReport,
+    &SubReport,
+    L"ACPI",
+    EFI_FILE_MODE_READ | EFI_FILE_MODE_WRITE | EFI_FILE_MODE_CREATE,
+    EFI_FILE_DIRECTORY
+    );
+  if (!EFI_ERROR (Status)) {
+    DEBUG ((DEBUG_INFO, "OC: Dumping ACPI for report...\n"));
+    Status = AcpiDumpTables (SubReport);
+    SubReport->Close (SubReport);
+  }
+  DEBUG ((DEBUG_INFO, "OC: ACPI dumping - %r\n", Status)); 
+
+  Status = SafeFileOpen (
+    SysReport,
+    &SubReport,
+    L"SMBIOS",
+    EFI_FILE_MODE_READ | EFI_FILE_MODE_WRITE | EFI_FILE_MODE_CREATE,
+    EFI_FILE_DIRECTORY
+    );
+  if (!EFI_ERROR (Status)) {
+    DEBUG ((DEBUG_INFO, "OC: Dumping SMBIOS for report...\n"));
+    Status = OcSmbiosDump (SubReport);
+    SubReport->Close (SubReport);
+  }
+  DEBUG ((DEBUG_INFO, "OC: ACPI dumping - %r\n", Status)); 
+
+  SysReport->Close (SysReport);
+  Fs->Close (Fs);
+
+  return EFI_SUCCESS;
 }
 
 STATIC
@@ -539,6 +627,12 @@ OcMiscMiddleInit (
     );
 
   *LoadHandle = OcHandle;
+
+  DEBUG_CODE_BEGIN ();
+  if (OcHandle != NULL && Config->Misc.Debug.SysReport) {
+    ProduceDebugReport (OcHandle);
+  }
+  DEBUG_CODE_END ();
 
   return Status;
 }